Add custom Item marshaling
This PR adds the following interfaces and integrates them
into `MarshalItem()` and `UnmarshalItem()`.

They allow the complete customisation of item marshalling and
unmarshalling. This is particularly useful when one of the
attributes on your item is an `interface{}` as you need to
dynamically unmarshal/marshal to several different types.

``` go
// ItemUnmarshaler is the interface implemented by
// objects that can unmarshal an Item (a map of strings
//to AttributeValues) into themselves.
type ItemUnmarshaler interface {
	UnmarshalDynamoItem(item map[string]*dynamodb.AttributeValue) error
}

// ItemMarshaler is the interface implemented by objects that
// can marshal themselves into an Item (a map of strings
// to AttributeValues).
type ItemMarshaler interface {
	MarshalDynamoItem() (map[string]*dynamodb.AttributeValue, error)
}
```